refactor: packages folder be 2 (#1628)
# Description

Move some folders to packages
Remove unused code
