+++ backend/middlewares/auth/auth_bearer.py
@@ -1,10 +1,10 @@
 import os
 from typing import Optional
 
-from auth.api_key_handler import get_user_from_api_key, verify_api_key
-from auth.jwt_token_handler import decode_access_token, verify_token
 from fastapi import Depends, HTTPException, Request
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
+from middlewares.auth.api_key_handler import get_user_from_api_key, verify_api_key
+from middlewares.auth.jwt_token_handler import decode_access_token, verify_token
 from models import UserIdentity
