+++ backend/routes/crawl_routes.py
@@ -1,15 +1,15 @@
 from typing import Optional
 from uuid import UUID
 
-from auth import AuthBearer, get_current_user
 from celery_worker import process_crawl_and_notify
-from crawl.crawler import CrawlWebsite
 from fastapi import APIRouter, Depends, Query, Request
 from logger import get_logger
+from middlewares.auth import AuthBearer, get_current_user
 from models import Brain, UserIdentity, UserUsage
 from models.databases.supabase.knowledge import CreateKnowledgeProperties
 from models.databases.supabase.notifications import CreateNotificationProperties
 from models.notifications import NotificationsStatusEnum
+from packages.files.crawl.crawler import CrawlWebsite
 from packages.files.file import convert_bytes
 from repository.knowledge.add_knowledge import add_knowledge
 from repository.notification.add_notification import add_notification
