+++ backend/routes/knowledge_routes.py
@@ -1,15 +1,14 @@
 from uuid import UUID
 
-from auth import AuthBearer, get_current_user
 from fastapi import APIRouter, Depends, Query
 from logger import get_logger
+from middlewares.auth import AuthBearer, get_current_user
 from models import Brain, UserIdentity
 from repository.files.delete_file import delete_file_from_storage
 from repository.files.generate_file_signed_url import generate_file_signed_url
 from repository.knowledge.get_all_knowledge import get_all_knowledge
 from repository.knowledge.get_knowledge import get_knowledge
 from repository.knowledge.remove_knowledge import remove_knowledge
-
 from routes.authorizations.brain_authorization import (
     RoleEnum,
     has_brain_authorization,
