+++ backend/routes/user_routes.py
@@ -1,7 +1,7 @@
 import time
 
-from auth import AuthBearer, get_current_user
 from fastapi import APIRouter, Depends, Request
+from middlewares.auth import AuthBearer, get_current_user
 from models import Brain, UserIdentity, UserUsage
 from repository.brain import get_user_default_brain
 from repository.user_identity.get_user_identity import get_user_identity
